feat: add precision benchmark
